refactor(play-list-detail): 記述変更＆リネーム作業
