+++ src/app/create-list/create-list.component.html
@@ -34,7 +34,6 @@
     </form>
     <mat-form-field [formGroup]="form" class="create-list__selector">
       <mat-label>プライバシー設定</mat-label>
-      <mat-select-trigger> </mat-select-trigger>
       <mat-select formControlName="privacy" required>
         <mat-option value="public">
           公開
