+++ src/app/play-list-detail/list-description-edit/list-description-edit.component.html
@@ -1,11 +1,11 @@
 <ng-container *ngIf="playList$ | async as playList">
-  <div class="edit-explan">
-    <ng-container *ngIf="isTextEditable; else text">
-      <form [formGroup]="form" (ngSubmit)="updateListText()">
-        <mat-form-field class="edit-explan__form">
+  <div class="edit-desc">
+    <ng-container *ngIf="isEditable; else text">
+      <form [formGroup]="form" (ngSubmit)="updateDescription()">
+        <mat-form-field class="edit-desc__form">
           <textarea
-            formControlName="listText"
-            class="edit-explan__textarea"
+            formControlName="description"
+            class="edit-desc__textarea"
             autofocus
             matTextareaAutosize
             type="text"
@@ -13,10 +13,15 @@
             #input
             placeholder="説明"
           ></textarea>
-          <mat-hint align="end">{{ input.value.length }} / 5000</mat-hint>
-          <mat-error class="mat-error" *ngIf="listText.hasError('maxlength')">
-            <p>最大5000文字です</p>
-            <p>{{ input.value.length }} / 5000</p>
+          <mat-hint align="end"
+            >{{ input.value.length }} / {{ DescriptionMaxlength }}</mat-hint
+          >
+          <mat-error
+            class="mat-error"
+            *ngIf="description.hasError('maxlength')"
+          >
+            <p>最大{{ DescriptionMaxlength }}文字です</p>
+            <p>{{ input.value.length }} / {{ DescriptionMaxlength }}</p>
           </mat-error>
         </mat-form-field>
         <div class="edit-actions">
@@ -24,7 +29,7 @@
             class="edit-actions__btn"
             mat-button
             type="button"
-            (click)="isTextEditable = false; setValue()"
+            (click)="isEditable = false; setValue()"
           >
             キャンセル
           </button>
@@ -41,10 +46,10 @@
     </ng-container>
 
     <ng-template #text>
-      <p class="edit-explan__title">説明</p>
-      <div class="edit-explan__text">
-        <p>{{ playList.listText }}</p>
-        <button mat-icon-button (click)="isTextEditable = true">
+      <p class="edit-desc__title">説明</p>
+      <div class="edit-desc__text">
+        <p>{{ playList.description }}</p>
+        <button mat-icon-button (click)="isEditable = true">
           <mat-icon>create</mat-icon>
         </button>
       </div>
