+++ src/app/play-list-detail/list-description-edit/list-description-edit.component.scss
@@ -1,4 +1,4 @@
-.edit-explan {
+.edit-desc {
   padding-bottom: 16px;
   &__form {
     width: 100%;
