+++ src/app/play-list-detail/list-privacy-edit/list-privacy-edit.component.html
@@ -1,6 +1,5 @@
 <mat-form-field [formGroup]="form">
   <mat-label>プライバシー設定</mat-label>
-  <mat-select-trigger> </mat-select-trigger>
   <mat-select formControlName="privacy" required>
     <mat-option (click)="updatePrivacy()" value="public">
       公開
