+++ src/app/play-list-detail/list-privacy-edit/list-privacy-edit.component.ts
@@ -1,11 +1,10 @@
-import { Component, OnInit } from '@angular/core';
+import { Component, Input, OnInit } from '@angular/core';
+import { FormBuilder, FormControl, Validators } from '@angular/forms';
 import { ActivatedRoute } from '@angular/router';
-import { PlayListService } from 'src/app/services/play-list.service';
-import { Observable } from 'rxjs';
 import { PlayList } from 'functions/src/intarfaces/play-list';
-import { switchMap, map } from 'rxjs/operators';
+import { Observable } from 'rxjs';
 import { AuthService } from 'src/app/services/auth.service';
-import { FormBuilder, Validators, FormControl } from '@angular/forms';
+import { PlayListService } from 'src/app/services/play-list.service';
 
 @Component({
   selector: 'app-list-privacy-edit',
@@ -13,16 +12,11 @@ import { FormBuilder, Validators, FormControl } from '@angular/forms';
   styleUrls: ['./list-privacy-edit.component.scss'],
 })
 export class ListPrivacyEditComponent implements OnInit {
+  @Input() playList$: Observable<PlayList>;
+
   private uid = this.authService.uid;
-  private listId: string;
-  private listId$: Observable<string> = this.route.paramMap.pipe(
-    map((paramMap) => paramMap.get('id'))
-  );
 
   isMovieEditable: boolean;
-  playList$: Observable<PlayList> = this.listId$.pipe(
-    switchMap((listId) => this.playListService.getMyPlayList(this.uid, listId))
-  );
   form = this.fb.group({
     privacy: [
       '',
@@ -38,18 +32,10 @@ export class ListPrivacyEditComponent implements OnInit {
     private playListService: PlayListService,
     private authService: AuthService,
     private fb: FormBuilder
-  ) {
-    this.getListId();
-  }
+  ) {}
 
   ngOnInit(): void {
     this.setValue();
-  }
-
-  private getListId() {
-    this.listId$.subscribe((id) => {
-      this.listId = id;
-    });
   }
 
   private setValue() {
@@ -60,9 +46,10 @@ export class ListPrivacyEditComponent implements OnInit {
 
   updatePrivacy() {
     const formData = this.form.value;
+    const listId = this.route.snapshot.paramMap.get('id');
     const newValue = {
       privacy: formData.privacy,
     };
-    this.playListService.updatePlayList(this.uid, this.listId, newValue);
+    this.playListService.updatePlayList(this.uid, listId, newValue);
   }
 }
