+++ src/app/play-list-detail/video-addition/video-addition.component.html
@@ -5,6 +5,7 @@
         class="edit-movie__input"
         placeholder="動画のURLを入力して下さい"
         autocomplete="off"
+        type="url"
         matInput
         autofocus
         #input
